backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.20
langchain-google-genai>=1.0.3
langchain-openai>=0.1.3
pandas>=2.2.0
numpy>=1.26.3
yfinance>=0.2.35
python-dotenv>=1.0.0
pydantic>=2.5.3
httpx>=0.26.0
duckduckgo-search>=6.2.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pypdf2>=3.0.0
langchain-text-splitters>=0.0.1
python-multipart>=0.0.6
